# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SwimManager Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meets:
    get:
      operationId: listMeets
      tags: [meets]
      summary: List all meets
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of meets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meet"
    post:
      operationId: createMeet
      tags: [meets]
      summary: Create a meet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meet"
  /meets/{id}:
    get:
      operationId: getMeet
      tags: [meets]
      summary: Get meet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meet"
    patch:
      operationId: updateMeet
      tags: [meets]
      summary: Update meet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meet"
    delete:
      operationId: deleteMeet
      tags: [meets]
      summary: Delete meet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /meets/{id}/summary:
    get:
      operationId: getMeetSummary
      tags: [meets]
      summary: Get meet summary stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meet summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetSummary"
  /meets/{id}/team-scores:
    get:
      operationId: getMeetTeamScores
      tags: [meets]
      summary: Get team scores for a meet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamScore"

  # ── SESSIONS ──────────────────────────────────────────────────────────────
...
  /events/{eventId}/heats:
    get:
      operationId: listHeats
      tags: [heats]
      summary: List heats for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Heats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Heat"
    post:
      operationId: seedEvent
      tags: [heats]
      summary: Seed an event (generate heats)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedInput"
      responses:
        "201":
          description: Seeded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Heat"

  # ── RESULTS ───────────────────────────────────────────────────────────────
  /entries/{entryId}/result:
    get:
      operationId: getResult
      tags: [results]
      summary: Get result for an entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
    put:
      operationId: setResult
      tags: [results]
      summary: Set or update result for an entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /events/{eventId}/results:
    get:
      operationId: listEventResults
      tags: [results]
      summary: Get all results for an event sorted by place
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"

  # ── RELAYS ────────────────────────────────────────────────────────────────
  /events/{eventId}/relays:
    get:
      operationId: listRelays
      tags: [relays]
      summary: List relay entries for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Relay entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelayEntry"
    post:
      operationId: createRelay
      tags: [relays]
      summary: Create relay entry
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelayEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelayEntry"
  /relays/{id}:
    patch:
      operationId: updateRelay
      tags: [relays]
      summary: Update relay entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelayEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelayEntry"
    delete:
      operationId: deleteRelay
      tags: [relays]
      summary: Delete relay entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── TIME STANDARDS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Meet:
      type: object
      required: [id, name, startDate]
      properties:
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        facility:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        course:
          type: string
        meetType:
          type: string
        meetStyle:
          type: string
        meetClass:
          type: ["string", "null"]
        idFormat:
          type: ["string", "null"]
        hostLsc:
          type: ["string", "null"]
        altitude:
          type: ["integer", "null"]
        entryDeadline:
          type: ["string", "null"]
        ageUpDate:
          type: ["string", "null"]
        status:
          type: string
        scoringRules:
          type: ["string", "null"]
        lanes:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    MeetInput:
      type: object
      required: [name, startDate, course, meetType, meetStyle]
      properties:
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        facility:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        course:
          type: string
        meetType:
          type: string
        meetStyle:
          type: string
        meetClass:
          type: string
        idFormat:
          type: string
        hostLsc:
          type: string
        altitude:
          type: integer
        entryDeadline:
          type: string
        ageUpDate:
          type: string
        scoringRules:
          type: string
        lanes:
          type: integer
        notes:
          type: string
    MeetUpdate:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        facility:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        course:
          type: string
        meetType:
          type: string
        meetStyle:
          type: string
        meetClass:
          type: string
        idFormat:
          type: string
        hostLsc:
          type: string
        altitude:
          type: integer
        entryDeadline:
          type: string
        ageUpDate:
          type: string
        status:
          type: string
        scoringRules:
          type: string
        lanes:
          type: integer
        notes:
          type: string
    MeetSummary:
      type: object
      required: [meetId, totalAthletes, totalEvents, totalEntries, totalTeams, completedEvents, pendingEvents]
      properties:
        meetId:
          type: integer
        totalAthletes:
          type: integer
        totalEvents:
          type: integer
        totalEntries:
          type: integer
        totalTeams:
          type: integer
        completedEvents:
          type: integer
        pendingEvents:
          type: integer
        totalHeats:
          type: integer
    TeamScore:
      type: object
      required: [teamId, teamName, points, place]
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        teamAbbreviation:
          type: ["string", "null"]
        points:
          type: number
        place:
          type: integer

    # ── SESSION ───────────────────────────────────────────────────────────
...
    Heat:
      type: object
      required: [id, eventId, heatNumber]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        heatNumber:
          type: integer
        status:
          type: string
        lanes:
          type: array
          items:
            $ref: "#/components/schemas/HeatLane"
    HeatLane:
      type: object
      required: [lane]
      properties:
        lane:
          type: integer
        entryId:
          type: ["integer", "null"]
        athleteId:
          type: ["integer", "null"]
        athleteName:
          type: ["string", "null"]
        teamName:
          type: ["string", "null"]
        seedTime:
          type: ["number", "null"]
        finishTime:
          type: ["number", "null"]
        place:
          type: ["integer", "null"]
        dq:
          type: boolean
        ns:
          type: boolean
        dnf:
          type: boolean
    SeedInput:
      type: object
      required: [lanes, heatOrder]
      properties:
        lanes:
          type: integer
        heatOrder:
          type: string
        circleSeeding:
          type: boolean

    # ── RESULT ────────────────────────────────────────────────────────────
    Result:
      type: object
      required: [id, entryId, eventId]
      properties:
        id:
          type: integer
        entryId:
          type: integer
        eventId:
          type: integer
        athleteId:
          type: ["integer", "null"]
        athleteName:
          type: ["string", "null"]
        teamName:
          type: ["string", "null"]
        finishTime:
          type: ["number", "null"]
        place:
          type: ["integer", "null"]
        points:
          type: ["number", "null"]
        dq:
          type: boolean
        dqCode:
          type: ["string", "null"]
        ns:
          type: boolean
        dnf:
          type: boolean
        splits:
          type: ["string", "null"]
        heatNumber:
          type: ["integer", "null"]
        lane:
          type: ["integer", "null"]
    ResultInput:
      type: object
      properties:
        finishTime:
          type: number
        dq:
          type: boolean
        dqCode:
          type: string
        ns:
          type: boolean
        dnf:
          type: boolean
        splits:
          type: string
        place:
          type: integer
        points:
          type: number

    # ── RELAY ─────────────────────────────────────────────────────────────
    RelayEntry:
      type: object
      required: [id, eventId, teamId]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: ["string", "null"]
        seedTime:
          type: ["number", "null"]
        heatNumber:
          type: ["integer", "null"]
        lane:
          type: ["integer", "null"]
        scratched:
          type: boolean
        leg1AthleteId:
          type: ["integer", "null"]
        leg1AthleteName:
          type: ["string", "null"]
        leg2AthleteId:
          type: ["integer", "null"]
        leg2AthleteName:
          type: ["string", "null"]
        leg3AthleteId:
          type: ["integer", "null"]
        leg3AthleteName:
          type: ["string", "null"]
        leg4AthleteId:
          type: ["integer", "null"]
        leg4AthleteName:
          type: ["string", "null"]
        finishTime:
          type: ["number", "null"]
        place:
          type: ["integer", "null"]
    RelayEntryInput:
      type: object
      required: [teamId]
      properties:
        teamId:
          type: integer
        seedTime:
          type: number
        leg1AthleteId:
          type: integer
        leg2AthleteId:
          type: integer
        leg3AthleteId:
          type: integer
        leg4AthleteId:
          type: integer
    RelayEntryUpdate:
      type: object
      properties:
        seedTime:
          type: number
        heatNumber:
          type: integer
        lane:
          type: integer
        scratched:
          type: boolean
        leg1AthleteId:
          type: integer
        leg2AthleteId:
          type: integer
        leg3AthleteId:
          type: integer
        leg4AthleteId:
          type: integer
        finishTime:
          type: number
        place:
          type: integer

    # ── TIME STANDARDS ────────────────────────────────────────────────────
...