# lib/api-zod/src/generated/types/seedInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwimManager Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface SeedInput {
  lanes: number;
  heatOrder: string;
  circleSeeding?: boolean;
}